+++ packages/amplify-schema-validator/CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [0.1.1-beta.0](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-schema-validation@0.1.0...@aws-amplify/graphql-schema-validation@0.1.1-beta.0) (2023-02-03)
+## [0.1.1](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-schema-validation@0.1.0...@aws-amplify/graphql-schema-validation@0.1.1) (2023-02-10)
 
 ### Bug Fixes